requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sentence-transformers
faiss-cpu
pymongo
python-dotenv
pypdf
streamlit
requests
numpy
